src/utilities/gcp-documentAI-processor-module/requirements.txt
google-cloud-documentai>=2.16.0
google-api-core>=2.10.0
google-auth>=2.6.0
google-cloud-core>=2.3.0
requests>=2.27.1
setuptools>=65.5.1
protobuf>=3.19.0
google-cloud-storage>=2.0.0 